requirements.txt
openai==0.27.0
pandas
python-docx==0.8.11
streamlit==1.11.0
streamlit-chat==0.0.2.1
transformers==4.26.0
